+++ packages/forma-36-react-components/src/components/Typography/DisplayText/DisplayText.stories.tsx
@@ -1,14 +1,18 @@
 import React from 'react';
 import DisplayText, { DisplayTextProps } from './DisplayText';
 
+import Flex from '../../Flex/Flex';
+import Paragraph from './../Paragraph';
+
 export default {
-  title: 'Components/Typography/DisplayText',
+  title: 'Typography/DisplayText',
   component: DisplayText,
   parameters: {
     propTypes: [DisplayText['__docgenInfo']],
   },
   argTypes: {
     className: { control: { disable: true } },
+    testId: { control: { disable: true } },
   },
 };
 
@@ -21,7 +25,40 @@ export const Basic = ({ displayText, ...args }: Args) => (
 );
 
 Basic.args = {
-  element: 'h4',
-  size: 'default',
+  displayText: 'Display text',
+};
+
+export const Overview = ({ displayText, ...args }: Args) => (
+  <>
+    <Flex alignItems="center" marginBottom="spacingL">
+      <Flex marginRight="spacingS">
+        <Paragraph>48</Paragraph>
+      </Flex>
+      <DisplayText {...args} size="huge">
+        {displayText}
+      </DisplayText>
+    </Flex>
+
+    <Flex alignItems="center" marginBottom="spacingL">
+      <Flex marginRight="spacingS">
+        <Paragraph>36</Paragraph>
+      </Flex>
+      <DisplayText {...args} size="large">
+        {displayText}
+      </DisplayText>
+    </Flex>
+
+    <Flex alignItems="center">
+      <Flex marginRight="spacingS">
+        <Paragraph>28</Paragraph>
+      </Flex>
+      <DisplayText {...args} size="default">
+        {displayText}
+      </DisplayText>
+    </Flex>
+  </>
+);
+
+Overview.args = {
   displayText: 'Display text',
 };