+++ packages/forma-36-tokens/src/tokens/typography/font-size.js
@@ -1,7 +1,8 @@
 const fontSize = {
-  'font-size-3xl': '2.1875rem',
+  'font-size-4xl': '3rem',
+  'font-size-3xl': '2.25rem',
   'font-size-2xl': '1.75rem',
-  'font-size-xl': '1.3125rem',
+  'font-size-xl': '1.25rem',
   'font-size-l': '1rem',
   'font-size-m': '0.875rem',
   'font-size-s': '0.75rem',